Solutions for day 2.
Added `Serapeum` and `Alexandria` (hoped that they contained
something like `group-by`; they didn't or I didn't find it...)

Added the input to day 2 and the test cases that were given
in the puzzle.

The solution to part 2 is a brute force solution. It blindly
iterates through the input in quadratic fashion to find the first
pair that differs by exactly one character.